Replace parseInt -> parseFloat for margin-based examples

+++ src/comparisons/elements/outer_width_with_margin/ie9.js
@@ -2,7 +2,7 @@ function outerWidth(el) {
   var width = el.offsetWidth;
   var style = getComputedStyle(el);
 
-  width += parseInt(style.marginLeft) + parseInt(style.marginRight);
+  width += parseFloat(style.marginLeft) + parseFloat(style.marginRight);
   return width;
 }
